Consolidating types into one file, fixing type, simplifying some code
