--- src/altinn-app-frontend/src/features/form/dynamics/conditionalRendering/conditionalRenderingActions.ts
@@ -1,3 +1,0 @@
-export interface ICheckIfConditionalRulesShouldRun {
-  repeatingContainerId?: string;
-}
