--- src/altinn-app-frontend/src/features/form/dynamics/fetch/fetchFormDynamicsActions.ts
@@ -1,9 +1,0 @@
-export interface IFetchServiceConfigFulfilled {
-  apis: any;
-  ruleConnection: any;
-  conditionalRendering: any;
-}
-
-export interface IFetchServiceConfigRejected {
-  error: Error;
-}
