+++ src/altinn-app-frontend/src/features/form/dynamics/index.ts
@@ -1,5 +1,3 @@
-import type { IConditionalRenderingRules } from './types';
-
 export interface IFormDynamicState {
   apis: any;
   ruleConnection: IRuleConnections;
@@ -18,4 +16,43 @@ export interface IRuleConnection {
 
 export interface IRuleConnections {
   [id: string]: IRuleConnection;
+}
+
+export interface ICheckIfConditionalRulesShouldRun {
+  repeatingContainerId?: string;
+}
+
+export interface IFetchServiceConfigFulfilled {
+  apis: any;
+  ruleConnection: any;
+  conditionalRendering: any;
+}
+
+export interface IFetchServiceConfigRejected {
+  error: Error;
+}
+
+export interface IConditionalRenderingRules {
+  [id: string]: IConditionalRenderingRule;
+}
+
+export interface IConditionalRenderingRule {
+  selectedFunction: string;
+  selectedAction: string;
+  selectedFields: ISelectedFields;
+  inputParams: IParameters;
+  repeatingGroup?: IConditionalRenderingRepeatingGroup;
+}
+
+export interface IConditionalRenderingRepeatingGroup {
+  groupId: string;
+  childGroupId?: string;
+}
+
+export interface IParameters {
+  [id: string]: string;
+}
+
+export interface ISelectedFields {
+  [id: string]: string;
 }
