--- src/altinn-app-frontend/src/features/form/dynamics/types/index.ts
@@ -1,24 +1,0 @@
-export interface IConditionalRenderingRules {
-  [id: string]: IConditionalRenderingRule;
-}
-
-export interface IConditionalRenderingRule {
-  selectedFunction: string;
-  selectedAction: string;
-  selectedFields: ISelectedFields;
-  inputParams: IParameters;
-  repeatingGroup?: IConditionalRenderingRepeatingGroup;
-}
-
-export interface IConditionalRenderingRepeatingGroup {
-  groupId: string;
-  childGroupId?: string;
-}
-
-export interface IParameters {
-  [id: string]: string;
-}
-
-export interface ISelectedFields {
-  [id: string]: string;
-}
